requirements.txt
requests
python-dotenv
PyYAML
Jinja2